# backend/src/swagger.json
{
    "openapi": "3.0.0",
    "info": {
      "title": "User API",
      "version": "1.0.0",
      "description": "API pour gérer les utilisateurs"
    },
    "paths": {
      "/users": {
        "get": {
          "summary": "Récupérer tous les utilisateurs",
          "responses": {
            "200": {
              "description": "Liste des utilisateurs"
            }
          }
        },
        "post": {
          "summary": "Créer un nouvel utilisateur",
          "requestBody": {
            "required": true,
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "nom": {
                      "type": "string"
                    },
                    "prenom": {
                      "type": "string"
                    },
                    "date_naissance": {
                      "type": "string",
                      "format": "date"
                    },
                    "genre": {
                      "type": "string"
                    },
                    "mail": {
                      "type": "string"
                    },
                    "password": {
                      "type": "string"
                    },
                    "telephone": {
                      "type": "string"
                    },
                    "actif": {
                      "type": "integer"
                    }
                  }
                }
              }
            },
            "responses": {
              "201": {
                "description": "Utilisateur créé"
              }
            }
          }
        }
      },
      "/users/{id}": {
        "get": {
          "summary": "Récupérer un utilisateur par ID",
          "parameters": [
            {
              "name": "id",
              "in": "path",
              "required": true,
              "schema": {
                "type": "integer"
              }
            }
          ],
          "responses": {
            "200": {
              "description": "Détails de l'utilisateur"
            }
          }
        },
        "put": {
          "summary": "Mettre à jour un utilisateur",
          "parameters": [
            {
              "name": "id",
              "in": "path",
              "required": true,
              "schema": {
                "type": "integer"
              }
            }
          ],
          "requestBody": {
            "required": true,
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "nom": {
                      "type": "string"
                    },
                    "prenom": {
                      "type": "string"
                    },
                    "date_naissance": {
                      "type": "string",
                      "format": "date"
                    },
                    "genre": {
                      "type": "string"
                    },
                    "mail": {
                      "type": "string"
                    },
                    "password": {
                      "type": "string"
                    },
                    "telephone": {
                      "type": "string"
                    },
                    "actif": {
                      "type": "integer"
                    }
                  }
                }
              }
            },
            "responses": {
              "200": {
                "description": "Utilisateur mis à jour"
              }
            }
          }
        },
        "delete": {
          "summary": "Supprimer un utilisateur",
          "parameters": [
            {
              "name": "id",
              "in": "path",
              "required": true,
              "schema": {
                "type": "integer"
              }
            }
          ],
          "responses": {
            "204": {
              "description": "Utilisateur supprimé"
            }
          }
        }
      }
    }
  }
  